requirements.txt
flask
sqlalchemy
pymysql
faiss-cpu
sentence-transformers
numpy
requests
python-dotenv
